# ArriveLocal-BE/GOOGLE_API_CONTRACT.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "ArriveLocal Hyperlocal API",
      "version": "1.0.0",
      "description": "API contract for ArriveLocal Hyperlocal Inventory + Delivery System"
    },
    "servers": [
      {
        "url": "http://localhost:5000/api",
        "description": "Local Dev Server"
      }
    ],
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Register new user or seller",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": { "type": "string" },
                    "email": { "type": "string" },
                    "password": { "type": "string" },
                    "role": { "type": "string", "enum": ["user", "seller", "delivery_partner"] }
                  }
                }
              }
            }
          },
          "responses": {
            "201": { "description": "User registered successfully" }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Login user/seller/delivery partner",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "email": { "type": "string" },
                    "password": { "type": "string" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Login successful with JWT token" }
          }
        }
      },
...
    }
  }